fix: add capsule client check before rendering provider-ui components inside examples
