+++ example/starter/src/pages/index.tsx
@@ -1,19 +1,9 @@
-import { Stack, useColorMode } from "@chakra-ui/react";
-import { useCapsule } from "graz";
-import dynamic from "next/dynamic";
+import { Stack } from "@chakra-ui/react";
 import { Card } from "src/ui/card/chain";
+import CapsuleModals from "src/ui/modal/capsule-modals";
 import { mainnetChains } from "src/utils/graz";
-import "@leapwallet/cosmos-social-login-capsule-provider-ui/styles.css";
-
-const LeapSocialLogin = dynamic(
-  () => import("@leapwallet/cosmos-social-login-capsule-provider-ui").then((m) => m.CustomCapsuleModalView),
-  { ssr: false },
-);
 
 const HomePage = () => {
-  const { client, modalState, onAfterLoginSuccessful, setModalState, onLoginFailure } = useCapsule();
-
-  const { colorMode } = useColorMode();
   return (
     <>
       <Stack spacing={4}>
@@ -21,22 +11,7 @@ const HomePage = () => {
           <Card key={chain.chainId} chain={chain} />
         ))}
       </Stack>
-      <div className='leap-ui'>
-        <LeapSocialLogin
-          capsule={client?.getClient() || undefined}
-          // @ts-expect-error - type error
-          oAuthMethods={["GOOGLE", "FACEBOOK", "TWITTER", "DISCORD", "APPLE"]}
-          onAfterLoginSuccessful={() => {
-            void onAfterLoginSuccessful?.();
-          }}
-          onLoginFailure={() => {
-            onLoginFailure();
-          }}
-          setShowCapsuleModal={setModalState}
-          showCapsuleModal={modalState}
-          theme={colorMode}
-        />
-      </div>
+      <CapsuleModals />
     </>
   );
 };
