+++ example/vite/src/App.tsx
@@ -3,6 +3,7 @@ import "./App.css";
 import { useAccount, useActiveChainIds, useConnect, useDisconnect } from "graz";
 
 import reactLogo from "./assets/react.svg";
+import CapsuleModals from "./components/CapsuleModals";
 
 // eslint-disable-next-line prefer-arrow-functions/prefer-arrow-functions, react/function-component-definition
 export default function App() {
@@ -47,6 +48,7 @@ export default function App() {
           {isDisconnected ? "Connect Wallet" : null}
         </button>
       </div>
+      <CapsuleModals />
     </div>
   );
 }
